Change int id to long id in Author

+++ src/main/java/ru/dsoccer1980/dao/AuthorDao.java
@@ -10,12 +10,12 @@ public interface AuthorDao {
 
     void insert(String name);
 
-    Author getById(int id);
+    Author getById(long id);
 
     List<Author> getAll();
 
-    void deleteById(int id);
+    void deleteById(long id);
 
-    int getIdByName(String name);
+    long getIdByName(String name);
 
 }
